bot/requirements.txt
aiogram==3.0.0
anyio
Jinja2==3.1.2
python-dotenv==1.0.0
SQLAlchemy==2.0.20
PyMySQL==1.1.0